chore: update 'eventual-send' package to use AVA
This includes a temporary workaround in test-hp.js, @michaelfig will think
about what a better test should look like
